Split code into headers/source
