--- Box.h
@@ -1,8 +1,0 @@
-//
-// Created by Alihan Baysal on 12/2/25.
-//
-
-#ifndef SDL_GAMES_BOX_H
-#define SDL_GAMES_BOX_H
-
-#endif //SDL_GAMES_BOX_H
