+++ Headers/constants.h
@@ -1,0 +1,11 @@
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+#include <SDL3/SDL.h>
+
+
+constexpr int ScreenWidth {640};
+constexpr int ScreenHeight {480};
+constexpr int ScreenFps = 60;
+constexpr SDL_Color bgColor {0, 0, 0, 255};
+
+#endif
