+++ Headers/util.h
@@ -1,0 +1,7 @@
+#ifndef UTIL_H
+#define UTIL_H
+
+void close();
+bool init();
+
+#endif
