+++ Source/globals.cpp
@@ -1,0 +1,8 @@
+#include "../Headers/globals.h"
+
+SDL_Window* window {nullptr};
+SDL_Renderer* renderer {nullptr};
+int score {0};
+int hitScale {0};
+TTF_Font* Font;
+Texture scoreTexture;
