+++ Source/util.cpp
@@ -1,0 +1,34 @@
+#include "../Headers/util.h"
+#include "../Headers/globals.h"
+#include "../Headers/constants.h"
+#include <SDL3/SDL.h>
+
+
+void close() {
+    SDL_DestroyRenderer(renderer);
+    renderer = nullptr;
+    SDL_DestroyWindow(window);
+    window = nullptr;
+    SDL_Quit();
+}
+
+bool init() {
+    bool success {true};
+
+    if (!SDL_InitSubSystem(SDL_INIT_VIDEO | SDL_INIT_AUDIO)) {
+        SDL_Log("SDL could not be initialized! SDL error: %s\n", SDL_GetError());
+        success = false;
+    }
+
+    if (!SDL_CreateWindowAndRenderer("Pong", ScreenWidth, ScreenHeight, 0, &window, &renderer)) {
+        SDL_Log("SDL could not create window and renderer! SDL error: %s\n", SDL_GetError());
+        success = false;
+    }
+
+    if(!TTF_Init()) {
+        SDL_Log( "SDL_ttf could not initialize! SDL_ttf error: %s\n", SDL_GetError() );
+        success = false;
+    }
+
+    return success;
+}
